added label tests
