+++ package.json
@@ -12,15 +12,16 @@
     "electron-packager": "^17.1.2",
     "jest": "^29.7.0",
     "jsdoc": "^4.0.3",
+    "mock-fs": "^5.2.0",
     "prettier": "3.2.5",
     "wdio-electron-service": "^6.6.1"
   },
   "scripts": {
     "test": "node node_modules/jest/bin/jest.js -- __tests__/unit",
     "start": "electron .",
-    "package-win": "electron-packager . DevJournal --platform=win32 --arch=x64 --asar --out=DevJournal",
-    "package-mac": "electron-packager . DevJournal --platform=darwin --arch=x64 --asar --out=DevJournal",
-    "package-linux": "electron-packager . DevJournal --platform=linux --arch=x64 --asar --out=DevJournal",
+    "package-win": "electron-packager . DevJournal --platform=win32 --arch=x64 --asar --prune --out=DevJournal",
+    "package-mac": "electron-packager . DevJournal --platform=darwin --arch=x64 --asar --prune --out=DevJournal",
+    "package-linux": "electron-packager . DevJournal --platform=linux --arch=x64 --asar --prune --out=DevJournal",
     "wdio": "wdio run ./wdio.conf.js"
   },
   "jest": {
